fix: useNft hook 분리
